sibgrapi24-main/src/requirements.txt
h5py~=2.10.0
imutils~=0.5.4
ipykernel~=6.16.2
ipython~=7.34.0
ipython-genutils~=0.2.0
ISR~=2.2.0
kedro~=0.18.6
matplotlib~=3.5.3
numpy~=1.21.6
opencv-python~=4.10.0.84
pandas~=1.3.5
Pillow~=9.5.0
pytesseract~=0.3.10
scikit-image~=0.19.3
scikit-learn~=1.0.2
scipy~=1.7.3
segment-anything~=1.0
tensorflow~=2.0.0
torch~=1.13.1
torchmetrics~=0.11.4
torchvision~=0.14.1
pyenchant~=3.2.2